# user-service/apispec.yaml
...
openapi: 3.0.1
...
info:
  title: "User Service API"
  description: |
    User management microservice for Nexus collaborative supply chain platform.

    This service provides comprehensive user management capabilities including:
    - User account creation and management
    - User authentication and validation  
    - Batch user operations for performance optimization
    - User data retrieval and updates

    **Technical Stack:** Spring Boot 3.5.7, MongoDB, Java 21
    **Business Owner:** Product Engineering Department
    **Technical Contact:** Backend Engineering Team
  contact:
    name: "Backend Engineering Team"
    email: "backend-support@nexus.com"
  version: "1.0.0"

# =============================================================================
# SERVER CONFIGURATION
# =============================================================================
# Stakeholder: Infrastructure Team
# Purpose: Service endpoint configuration for different environments
servers:
  - url: "http://localhost:3000"
    description: "Generated server url"

# =============================================================================
# API TAGS & ORGANIZATION
# =============================================================================
# Stakeholder: API Design Team, Documentation Team
# Purpose: Logical grouping of endpoints for better organization
...
paths:
...
  /api/v1/users/{id}:
    
    # GET USER BY ID
    # Stakeholder: Customer Support Team, User Management Team
    # Use Case: Profile viewing, customer support inquiries
    get:
      tags:
        - "User Management"
      summary: "Get User by ID"
      description: "Retrieves detailed user information by user ID including wallet balance and funding requests. Business Stakeholder: Customer Support Team, Technical Owner: User Management Team, Use Case: Profile viewing and customer support"
      operationId: "getUserById"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "User retrieved successfully"
          content:
            "*/*":
              schema:
                type: "string"
        "400":
          description: "Invalid user ID format"
          content:
            "*/*":
              schema:
                type: "string"
        "404":
          description: "User not found"
          content:
            "*/*":
              schema:
                type: "string"
    
    # UPDATE USER
    # Stakeholder: Customer Operations Team, User Management Team
    # Use Case: Profile updates, wallet management, funding request updates
    put:
      tags:
        - "User Management"
      summary: "Update User Information"
      description: "Updates user profile information, wallet balance, and funding requests. Supports wallet adjustments and funding request management. Business Stakeholder: Customer Operations Team, Technical Owner: User Management Team, Use Case: Profile updates and wallet management"
      operationId: "updateUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequestDTO"
        required: true
      responses:
        "200":
          description: "User updated successfully"
          content:
            "*/*":
              schema:
                type: "string"
        "400":
          description: "Invalid input data or insufficient funds"
          content:
            "*/*":
              schema:
                type: "string"
        "404":
          description: "User not found"
          content:
            "*/*":
              schema:
                type: "string"
    
    # DELETE USER
    # Stakeholder: Data Management Team, User Management Team
    # Use Case: Account closure, data cleanup, GDPR compliance
    delete:
      tags:
        - "User Management"
      summary: "Delete User Account"
      description: "Permanently deletes a user account from the system. This action cannot be undone. Business Stakeholder: Data Management Team, Technical Owner: User Management Team, Use Case: Account closure and data cleanup"
      operationId: "deleteUser"
      parameters:
        - name: "id"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "User deleted successfully"
          content:
            "*/*":
              schema:
                type: "string"
        "400":
          description: "Invalid user ID format"
          content:
            "*/*":
              schema:
                type: "string"
        "404":
          description: "User not found"
          content:
            "*/*":
              schema:
                type: "string"
        "500":
          description: "Failed to delete user"
          content:
            "*/*":
              schema:
                type: "string"

  # ---------------------------------------------------------------------------
  # WALLET MANAGEMENT ENDPOINTS
  # ---------------------------------------------------------------------------
  # Business Context: Financial operations and wallet management
  # Primary Stakeholders: Finance Team, Customer Operations Team
  
  # WALLET WITHDRAW
  # Stakeholder: Finance Team, Security Operations Team
  # Use Case: Money withdrawal, financial transactions
  /api/v1/wallet/withdraw:
    post:
      tags:
        - "wallet-controller"
      operationId: "withdrawMoney"
      parameters:
        - name: "X-User-Id"
          in: "header"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletWithdrawRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            "*/*":
              schema:
                type: "object"
                additionalProperties:
                  type: "object"

  # WALLET DEPOSIT
  # Stakeholder: Finance Team, Customer Operations Team
  # Use Case: Money deposit, account funding
  /api/v1/wallet/deposit:
    post:
      tags:
        - "wallet-controller"
      operationId: "depositMoney"
      parameters:
        - name: "X-User-Id"
          in: "header"
          required: true
          schema:
            type: "string"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletDepositRequestDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            "*/*":
              schema:
                type: "object"
                additionalProperties:
                  type: "object"

  # ---------------------------------------------------------------------------
  # BATCH OPERATIONS ENDPOINT
  # ---------------------------------------------------------------------------
  # Business Context: Performance optimization for microservice communication
  # Primary Stakeholders: Integration Team, Performance Engineering Team
...
components:
  schemas:
...
    UserUpdateRequestDTO:
      type: "object"
      description: "Request DTO for updating user information including profile data, wallet balance, and funding requests"
      properties:
        name:
          type: "string"
          description: "User's full name"
          example: "John Doe"
        email:
          type: "string"
          description: "User's email address"
          example: "john.doe@example.com"
        walletBalance:
          type: "number"
          description: "New wallet balance (absolute value)"
          example: 1000.50
        walletAdjustment:
          type: "number"
          description: "Wallet balance adjustment (positive for credit, negative for debit)"
          example: -50.25
        fundingRequestIds:
          type: "array"
          description: "List of funding request IDs associated with the user"
          items:
            type: "string"
            example: "507f1f77bcf86cd799439011"

    # ---------------------------------------------------------------------------
    # WALLET TRANSACTION DTOs
    # ---------------------------------------------------------------------------
    # Stakeholder: Finance Team, Security Operations Team
    # Use Case: Financial transactions and wallet operations
    WalletWithdrawRequestDTO:
      type: "object"
      description: "Request DTO for wallet withdrawal operations"
      properties:
        amount:
          type: "number"
          description: "Amount to withdraw from wallet"
          example: 100.00
    WalletDepositRequestDTO:
      type: "object"
      description: "Request DTO for wallet deposit operations"
      properties:
        amount:
          type: "number"
          description: "Amount to deposit to wallet"
          example: 200.00

    # ---------------------------------------------------------------------------
    # BATCH USER REQUEST DTO
    # ---------------------------------------------------------------------------
    # Stakeholder: Integration Team, Performance Engineering Team
    # Use Case: Bulk user data retrieval for microservice communication
...